relocate utils.ts and walkthrough.ts test
