+++ packages/core/src/test/awsService/appBuilder/walkthrough.test.ts
@@ -9,22 +9,22 @@ import assert from 'assert'
 import {
     openApplicationComposerAfterReload,
     templateToOpenAppComposer,
-} from '../../../../awsService/appBuilder/activation'
-import globals from '../../../../shared/extensionGlobals'
+} from '../../../awsService/appBuilder/activation'
+import globals from '../../../shared/extensionGlobals'
 import {
     RuntimeLocationWizard,
     genWalkthroughProject,
     openProjectInWorkspace,
-} from '../../../../awsService/appBuilder/walkthrough'
-import { createWizardTester } from '../../../shared/wizards/wizardTestUtils'
-import { fs } from '../../../../shared'
-import { getTestWindow } from '../../../shared/vscode/window'
+} from '../../../awsService/appBuilder/walkthrough'
+import { createWizardTester } from '../../shared/wizards/wizardTestUtils'
+import { fs } from '../../../shared'
+import { getTestWindow } from '../../shared/vscode/window'
 
-import { AwsClis, installCli } from '../../../../shared/utilities/cliUtils'
-import { ChildProcess } from '../../../../shared/utilities/processUtils'
-import { assertTelemetryCurried } from '../../../testUtil'
-import { HttpResourceFetcher } from '../../../../shared/resourcefetcher/httpResourceFetcher'
-import { SamCliInfoInvocation } from '../../../../shared/sam/cli/samCliInfo'
+import { AwsClis, installCli } from '../../../shared/utilities/cliUtils'
+import { ChildProcess } from '../../../shared/utilities/processUtils'
+import { assertTelemetryCurried } from '../../testUtil'
+import { HttpResourceFetcher } from '../../../shared/resourcefetcher/httpResourceFetcher'
+import { SamCliInfoInvocation } from '../../../shared/sam/cli/samCliInfo'
 
 interface TestScenario {
     toolID: AwsClis
